requirements.txt
customtkinter
smolagents
pandas
requests
python-dotenv
litellm
